more refactoring
put all code parts into their respective files
